P4-3674: finally figure out a decent class overrides mechanism.
