+++ docker/customizations/any/temba/orgs/views.py
@@ -949,8 +949,8 @@ class UserCRUDL(SmartCRUDL):
             context["backup_tokens"] = self.get_user().backup_tokens.order_by("id")
             return context
 
-from engage.orgs.views import EngageOrgCRUDMixin
-class OrgCRUDL(EngageOrgCRUDMixin, SmartCRUDL):
+
+class OrgCRUDL(SmartCRUDL):
     actions = (
         "signup",
         "home",
