+++ engage/hamls/orgs/org_plan.haml
@@ -1,0 +1,7 @@
+-extends "smartmin/base.html"
+-load smartmin i18n humanize
+
+-block summary
+  .flex.items-center
+    .flex-grow.mr-4
+      You are currently on the '{{org.plan}}' plan, no topups will be required or tracked.
