+++ engage/orgs/views/manage_orgs.py
@@ -1,0 +1,78 @@
+import logging
+
+from django import forms
+from django.db.models import Sum
+from django.urls import reverse
+from django.utils.translation import gettext_lazy as _
+from django.views.generic.edit import CreateView
+
+from engage.utils.logs import LogExtrasMixin
+from engage.utils.strings import str2bool
+
+from temba.orgs.views import OrgCRUDL
+
+
+class AdminManageOverrides(LogExtrasMixin, OrgCRUDL.Manage):
+    mySuperType = OrgCRUDL.Manage
+    title = "Workspaces"
+    default_order = 'name'
+
+    def as_json(self, context):
+        pass
+
+    def get_gear_links(self) -> list:
+        links = [
+            dict(
+            #    id='btn-create-workspace',
+            #    title='Grant',
+            #    style='button-primary',
+            #    href=reverse("orgs.org_grant"),
+            #    as_btn=True,
+            #), dict(
+                id='btn-create-workspace',
+                title='New Workspace',
+                style='button-primary',
+                href=reverse("orgs.org_grant"),
+                as_btn=True,
+            ),
+        ]
+        return links
+    #enddef get_gear_links
+
+    def derive_queryset(self, **kwargs):
+        queryset = super(self.mySuperType, self).derive_queryset(**kwargs)
+        bActiveFilter = not self.request.GET.get("inactive")
+        queryset = queryset.filter(is_active=bActiveFilter)
+
+        brands = self.request.branding.get("keys")
+        if brands:
+            queryset = queryset.filter(brand__in=brands)
+
+        anon = self.request.GET.get("anon")
+        if anon:
+            queryset = queryset.filter(is_anon=str2bool(anon))
+
+        suspended = self.request.GET.get("suspended")
+        if suspended:
+            queryset = queryset.filter(is_suspended=str2bool(suspended))
+
+        flagged = self.request.GET.get("flagged")
+        if flagged:
+            queryset = queryset.filter(is_flagged=str2bool(flagged))
+
+        queryset = queryset.annotate(credits=Sum("topups__credits"))
+        queryset = queryset.annotate(paid=Sum("topups__price"))
+
+        return queryset
+    #enddef derive_queryset
+
+    def get_context_data(self, **kwargs):
+        context = super(self.mySuperType, self).get_context_data(**kwargs)
+        context["searches"] = []
+        context["anon_query"] = str2bool(self.request.GET.get("anon"))
+        context["flagged_query"] = str2bool(self.request.GET.get("flagged"))
+        context["suspended_query"] = str2bool(self.request.GET.get("suspended"))
+        return context
+    #enddef get_context_data
+
+#endclass AdminManageMixin
