+++ engage/utils/class_overrides.py
@@ -1,0 +1,18 @@
+from django.db.models import Model
+
+
+def ignoreDjangoModelAttrs(aDjangoModelClass: type[Model]):
+    """
+    Django models contain some class attributes we should avoid stomping when
+    we define MyClassOverrides class containers.
+    :param aDjangoModelClass: a model class we will be expanding/overriding.
+    :return: Returns a list of attributes to ignore when calling setClassOverrides().
+    """
+    return (
+        '_meta',
+        'DoesNotExist',
+        'MultipleObjectsReturned',
+        f"{aDjangoModelClass.__name__.lower()}_ptr_id",
+        f"{aDjangoModelClass.__name__.lower()}_ptr",
+    )
+#enddef ignoreDjangoModelAttrs
